requirements.txt
chromadb==1.2.1
huggingface-hub==0.35.3
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.79
langchain-huggingface==0.3.1
pandas==2.3.3
python-dotenv==1.1.1
sentence-transformers==5.1.1
